+++ GW2EIBuilders/JsonModels/JsonActors/JsonPlayerBuilder.cs
@@ -124,11 +124,11 @@ namespace GW2EIBuilders.JsonModels
             return jsonPlayer;
         }
 
-        private static List<JsonPlayerBuffsGeneration> GetPlayerBuffGenerations(List<IReadOnlyDictionary<long, FinalPlayerBuffs>> buffs, ParsedEvtcLog log, Dictionary<string, JsonLog.BuffDesc> buffDesc)
+        private static List<JsonPlayerBuffsGeneration> GetPlayerBuffGenerations(List<IReadOnlyDictionary<long, FinalActorBuffs>> buffs, ParsedEvtcLog log, Dictionary<string, JsonLog.BuffDesc> buffDesc)
         {
             IReadOnlyList<PhaseData> phases = log.FightData.GetNonDummyPhases(log);
             var uptimes = new List<JsonPlayerBuffsGeneration>();
-            foreach (KeyValuePair<long, FinalPlayerBuffs> pair in buffs[0])
+            foreach (KeyValuePair<long, FinalActorBuffs> pair in buffs[0])
             {
                 Buff buff = log.Buffs.BuffsByIds[pair.Key];
                 if (!buffDesc.ContainsKey("b" + pair.Key))
@@ -138,7 +138,7 @@ namespace GW2EIBuilders.JsonModels
                 var data = new List<JsonBuffsGenerationData>();
                 for (int i = 0; i < phases.Count; i++)
                 {
-                    if (buffs[i].TryGetValue(pair.Key, out FinalPlayerBuffs val))
+                    if (buffs[i].TryGetValue(pair.Key, out FinalActorBuffs val))
                     {
                         JsonBuffsGenerationData value = JsonPlayerBuffsGenerationBuilder.BuildJsonBuffsGenerationData(val);
                         data.Add(value);
@@ -165,12 +165,12 @@ namespace GW2EIBuilders.JsonModels
             return uptimes;
         }
 
-        private static List<JsonBuffsUptime> GetPlayerJsonBuffsUptime(Player player, List<IReadOnlyDictionary<long, FinalPlayerBuffs>> buffs, ParsedEvtcLog log, RawFormatSettings settings, Dictionary<string, JsonLog.BuffDesc> buffDesc, Dictionary<string, HashSet<long>> personalBuffs)
+        private static List<JsonBuffsUptime> GetPlayerJsonBuffsUptime(Player player, List<IReadOnlyDictionary<long, FinalActorBuffs>> buffs, ParsedEvtcLog log, RawFormatSettings settings, Dictionary<string, JsonLog.BuffDesc> buffDesc, Dictionary<string, HashSet<long>> personalBuffs)
         {
             var res = new List<JsonBuffsUptime>();
             var profEnums = new HashSet<ParserHelper.Source>(ProfToEnum(player.Prof));
             IReadOnlyList<PhaseData> phases = log.FightData.GetNonDummyPhases(log);
-            foreach (KeyValuePair<long, FinalPlayerBuffs> pair in buffs[0])
+            foreach (KeyValuePair<long, FinalActorBuffs> pair in buffs[0])
             {
                 Buff buff = log.Buffs.BuffsByIds[pair.Key];
                 var data = new List<JsonBuffsUptimeData>();
@@ -178,7 +178,7 @@ namespace GW2EIBuilders.JsonModels
                 {
                     PhaseData phase = phases[i];
                     Dictionary<long, FinalBuffsDictionary> buffsDictionary = player.GetBuffsDictionary(log, phase.Start, phase.End);
-                    if (buffs[i].TryGetValue(pair.Key, out FinalPlayerBuffs val))
+                    if (buffs[i].TryGetValue(pair.Key, out FinalActorBuffs val))
                     {
                         JsonBuffsUptimeData value = JsonBuffsUptimeBuilder.BuildJsonBuffsUptimeData(val, buffsDictionary[pair.Key]);
                         data.Add(value);
