+++ GW2EIEvtcParser/EIData/Actors/AbstractPlayer.cs
@@ -1,0 +1,233 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+using GW2EIEvtcParser.Exceptions;
+using GW2EIEvtcParser.ParsedData;
+using static GW2EIEvtcParser.EIData.Buff;
+using static GW2EIEvtcParser.ParserHelper;
+
+namespace GW2EIEvtcParser.EIData
+{
+    public class AbstractPlayer : AbstractSingleActor
+    {
+
+        private List<Consumable> _consumeList;
+        private List<DeathRecap> _deathRecaps;
+        // statistics
+        private CachingCollection<FinalPlayerSupport> _playerSupportStats;
+        //weaponslist
+        private string[] _weaponsArray;
+
+        // Constructors
+        internal AbstractPlayer(AgentItem agent) : base(agent)
+        {
+            if (agent.IsNPC)
+            {
+                throw new EvtcAgentException("Agent is NPC");
+            }
+            if (IsFakeActor)
+            {
+                throw new EvtcAgentException("Players can't be fake actors");
+            }
+        }
+
+        public override string GetIcon()
+        {
+            return AgentItem.Type == AgentItem.AgentType.NonSquadPlayer && !AgentItem.IsNotInSquadFriendlyPlayer ? GetHighResolutionProfIcon(Prof) : GetProfIcon(Prof);
+        }
+
+        // Public methods
+
+        public FinalPlayerSupport GetPlayerSupportStats(ParsedEvtcLog log, long start, long end)
+        {
+            if(_playerSupportStats == null)
+            {
+                _playerSupportStats = new CachingCollection<FinalPlayerSupport>(log);
+            }
+            if (!_playerSupportStats.TryGetValue(start, end, out FinalPlayerSupport value))
+            {
+                value = new FinalPlayerSupport(log, this, start, end);
+                _playerSupportStats.Set(start, end, value);
+            }
+            return value;
+        }
+
+        public IReadOnlyList<DeathRecap> GetDeathRecaps(ParsedEvtcLog log)
+        {
+            if (_deathRecaps == null)
+            {
+                SetDeathRecaps(log);
+            }
+            return _deathRecaps;
+        }
+
+        public IReadOnlyList<string> GetWeaponsArray(ParsedEvtcLog log)
+        {
+            if (_weaponsArray == null)
+            {
+                EstimateWeapons(log);
+            }
+            return _weaponsArray;
+        }
+
+        public IReadOnlyList<Consumable> GetConsumablesList(ParsedEvtcLog log, long start, long end)
+        {
+            if (_consumeList == null)
+            {
+                SetConsumablesList(log);
+            }
+            return _consumeList.Where(x => x.Time >= start && x.Time <= end).ToList();
+        }
+
+        // Private Methods
+
+        private void SetDeathRecaps(ParsedEvtcLog log)
+        {
+            _deathRecaps = new List<DeathRecap>();
+            IReadOnlyList<DeadEvent> deads = log.CombatData.GetDeadEvents(AgentItem);
+            IReadOnlyList<DownEvent> downs = log.CombatData.GetDownEvents(AgentItem);
+            IReadOnlyList<AliveEvent> ups = log.CombatData.GetAliveEvents(AgentItem);
+            long lastDeathTime = 0;
+            IReadOnlyList<AbstractHealthDamageEvent> damageLogs = GetDamageTakenEvents(null, log, 0, log.FightData.FightEnd);
+            foreach (DeadEvent dead in deads)
+            {
+                _deathRecaps.Add(new DeathRecap(log, damageLogs, dead, downs, ups, lastDeathTime));
+                lastDeathTime = dead.Time;
+            }
+        }
+
+        private void EstimateWeapons(ParsedEvtcLog log)
+        {
+            if (Prof == "Sword")
+            {
+                _weaponsArray = new string[]
+                {
+                    "Sword",
+                    "2Hand",
+                    null,
+                    null,
+                    null,
+                    null,
+                    null,
+                    null
+                };
+                return;
+            }
+            string[] weapons = new string[8];//first 2 for first set next 2 for second set, second sets of 4 for underwater
+            IReadOnlyList<AbstractCastEvent> casting = GetCastEvents(log, 0, log.FightData.FightEnd);
+            int swapped = -1;
+            long swappedTime = 0;
+            var swaps = casting.OfType<WeaponSwapEvent>().Select(x =>
+            {
+                return x.SwappedTo;
+            }).ToList();
+            foreach (AbstractCastEvent cl in casting)
+            {
+                if (cl.ActualDuration == 0 && cl.SkillId != SkillItem.WeaponSwapId)
+                {
+                    continue;
+                }
+                SkillItem skill = cl.Skill;
+                // first iteration
+                if (swapped == -1)
+                {
+                    swapped = skill.FindWeaponSlot(swaps);
+                }
+                if (!skill.EstimateWeapons(weapons, swapped, cl.Time > swappedTime + WeaponSwapDelayConstant) && cl is WeaponSwapEvent swe)
+                {
+                    //wepswap  
+                    swapped = swe.SwappedTo;
+                    swappedTime = swe.Time;
+                }
+            }
+            _weaponsArray = weapons;
+        }
+
+        private void SetConsumablesList(ParsedEvtcLog log)
+        {
+            IReadOnlyList<Buff> consumableList = log.Buffs.BuffsByNature[BuffNature.Consumable];
+            _consumeList = new List<Consumable>();
+            long fightDuration = log.FightData.FightEnd;
+            foreach (Buff consumable in consumableList)
+            {
+                foreach (AbstractBuffEvent c in log.CombatData.GetBuffData(consumable.ID))
+                {
+                    if (!(c is BuffApplyEvent ba) || AgentItem != ba.To)
+                    {
+                        continue;
+                    }
+                    long time = 0;
+                    if (!ba.Initial)
+                    {
+                        time = ba.Time;
+                    }
+                    if (time <= fightDuration)
+                    {
+                        Consumable existing = _consumeList.Find(x => x.Time == time && x.Buff.ID == consumable.ID);
+                        if (existing != null)
+                        {
+                            existing.Stack++;
+                        }
+                        else
+                        {
+                            _consumeList.Add(new Consumable(consumable, time, ba.AppliedDuration));
+                        }
+                    }
+                }
+            }
+            _consumeList.Sort((x, y) => x.Time.CompareTo(y.Time));
+
+        }
+
+        protected override void InitAdditionalCombatReplayData(ParsedEvtcLog log)
+        {
+            // Fight related stuff
+            log.FightData.Logic.ComputePlayerCombatReplayActors(this, log, CombatReplay);
+            if (CombatReplay.Rotations.Any())
+            {
+                CombatReplay.Decorations.Add(new FacingDecoration(((int)CombatReplay.TimeOffsets.start, (int)CombatReplay.TimeOffsets.end), new AgentConnector(this), CombatReplay.PolledRotations));
+            }
+        }
+
+        //
+
+        public override AbstractSingleActorSerializable GetCombatReplayJSON(CombatReplayMap map, ParsedEvtcLog log)
+        {
+            if (CombatReplay == null)
+            {
+                InitCombatReplay(log);
+            }
+            return new PlayerSerializable(this, log, map, CombatReplay);
+        }
+
+        public List<Point3D> GetCombatReplayActivePositions(ParsedEvtcLog log)
+        {
+            if (CombatReplay == null)
+            {
+                InitCombatReplay(log);
+            }
+            (IReadOnlyList<(long start, long end)> deads, _, IReadOnlyList<(long start, long end)> dcs) = GetStatus(log);
+            var activePositions = new List<Point3D>(GetCombatReplayPolledPositions(log));
+            for (int i = 0; i < activePositions.Count; i++)
+            {
+                Point3D cur = activePositions[i];
+                foreach ((long start, long end) in deads)
+                {
+                    if (cur.Time >= start && cur.Time <= end)
+                    {
+                        activePositions[i] = null;
+                    }
+                }
+                foreach ((long start, long end) in dcs)
+                {
+                    if (cur.Time >= start && cur.Time <= end)
+                    {
+                        activePositions[i] = null;
+                    }
+                }
+            }
+            return activePositions;
+        }
+    }
+}
