+++ GW2EIEvtcParser/EIData/Actors/AbstractSingleActor.cs
@@ -4,12 +4,15 @@ using System.Linq;
 using GW2EIEvtcParser.Exceptions;
 using GW2EIEvtcParser.ParsedData;
 using static GW2EIEvtcParser.EIData.Buff;
+using static GW2EIEvtcParser.ParserHelper;
 
 namespace GW2EIEvtcParser.EIData
 {
     public abstract class AbstractSingleActor : AbstractActor
     {
         public AgentItem AgentItem => SourceAgent;
+        public string Account { get; protected set; }
+        public int Group { get; protected set; }
         public int UniqueID => AgentItem.UniqueID;
         // Boons
         private HashSet<Buff> _trackedBuffs;
@@ -17,6 +20,7 @@ namespace GW2EIEvtcParser.EIData
         protected Dictionary<long, BuffsGraphModel> BuffGraphs { get; set; }
         private CachingCollection<BuffDistribution> _buffDistribution;
         private CachingCollection<Dictionary<long, long>> _buffPresence;
+        private CachingCollectionCustom<BuffEnum, Dictionary<long, FinalActorBuffs>[]> _buffStats;
         private CachingCollection<(Dictionary<long, FinalBuffsDictionary>, Dictionary<long, FinalBuffsDictionary>)> _buffsDictionary;
         private readonly Dictionary<long, AbstractBuffSimulator> _buffSimulators = new Dictionary<long, AbstractBuffSimulator>();
         // damage list
@@ -25,8 +29,12 @@ namespace GW2EIEvtcParser.EIData
         private CachingCollectionWithTarget<List<AbstractHealthDamageEvent>> _hitSelfDamageEventsPerPhasePerTarget;
         private CachingCollectionWithTarget<List<AbstractHealthDamageEvent>> _powerHitSelfDamageEventsPerPhasePerTarget;
         private CachingCollectionWithTarget<List<AbstractHealthDamageEvent>> _conditionHitSelfDamageEventsPerPhasePerTarget;
+        // Damage modifiers
+        private CachingCollectionWithTarget<Dictionary<string, DamageModifierStat>> _damageModifiersPerTargets;
+        private CachingCollectionWithTarget<Dictionary<string, List<DamageModifierEvent>>> _damageModifierEventsPerTargets;
         //status
         private List<Segment> _healthUpdates { get; set; }
+        private List<Segment> _breakbarPercentUpdates { get; set; }
         private List<Segment> _barrierUpdates { get; set; }
         private List<(long start, long end)> _deads;
         private List<(long start, long end)> _downs;
@@ -43,6 +51,8 @@ namespace GW2EIEvtcParser.EIData
 
         protected AbstractSingleActor(AgentItem agent) : base(agent)
         {
+            Group = 51;
+            Account = Character;
         }
 
         // Status
@@ -111,6 +121,15 @@ namespace GW2EIEvtcParser.EIData
                 _healthUpdates = Segment.FromStates(log.CombatData.GetHealthUpdateEvents(AgentItem).Select(x => x.ToState()).ToList(), 0, log.FightData.FightEnd);
             }
             return _healthUpdates;
+        }
+
+        public IReadOnlyList<Segment> GetBreakbarPercentUpdates(ParsedEvtcLog log)
+        {
+            if (_breakbarPercentUpdates == null)
+            {
+                _breakbarPercentUpdates = Segment.FromStates(log.CombatData.GetBreakbarPercentEvents(AgentItem).Select(x => x.ToState()).ToList(), 0, log.FightData.FightEnd);
+            }
+            return _breakbarPercentUpdates;
         }
 
         public IReadOnlyList<Segment> GetBarrierUpdates(ParsedEvtcLog log)
@@ -276,6 +295,100 @@ namespace GW2EIEvtcParser.EIData
             return brkDmgList;
         }
 
+        // Damage Modifiers
+        private Dictionary<string, DamageModifierStat> ComputeDamageModifierStats(NPC target, ParsedEvtcLog log, long start, long end)
+        {
+            // Check if damage mods against target
+            if (_damageModifierEventsPerTargets.TryGetValue(log.FightData.FightStart, log.FightData.FightEnd, target, out Dictionary<string, List<DamageModifierEvent>> events))
+            {
+                var res = new Dictionary<string, DamageModifierStat>();
+                foreach (KeyValuePair<string, List<DamageModifierEvent>> pair in events)
+                {
+                    DamageModifier damageMod = pair.Value.FirstOrDefault()?.DamageModifier;
+                    if (damageMod != null)
+                    {
+                        var eventsToUse = pair.Value.Where(x => x.Time >= start && x.Time <= end).ToList();
+                        int totalDamage = damageMod.GetTotalDamage(this, log, target, start, end);
+                        IReadOnlyList<AbstractHealthDamageEvent> typeHits = damageMod.GetHitDamageEvents(this, log, target, start, end);
+                        res[pair.Key] = new DamageModifierStat(eventsToUse.Count, typeHits.Count, eventsToUse.Sum(x => x.DamageGain), totalDamage);
+                    }
+                }
+                _damageModifiersPerTargets.Set(start, end, target, res);
+                return res;
+            }
+            // Check if we already filled the cache, that means no damage modifiers against given target
+            else if (_damageModifierEventsPerTargets.TryGetValue(log.FightData.FightStart, log.FightData.FightEnd, null, out events))
+            {
+                var res = new Dictionary<string, DamageModifierStat>();
+                _damageModifiersPerTargets.Set(start, end, target, res);
+                return res;
+            }
+            return null;
+        }
+
+        public IReadOnlyDictionary<string, DamageModifierStat> GetDamageModifierStats(NPC target, ParsedEvtcLog log, long start, long end)
+        {
+            if (!log.ParserSettings.ComputeDamageModifiers || IsDummyActor)
+            {
+                return new Dictionary<string, DamageModifierStat>();
+            }
+            if (_damageModifiersPerTargets == null)
+            {
+                _damageModifiersPerTargets = new CachingCollectionWithTarget<Dictionary<string, DamageModifierStat>>(log);
+                _damageModifierEventsPerTargets = new CachingCollectionWithTarget<Dictionary<string, List<DamageModifierEvent>>>(log);
+            }
+            if (_damageModifiersPerTargets.TryGetValue(start, end, target, out Dictionary<string, DamageModifierStat> res))
+            {
+                return res;
+            }
+            res = ComputeDamageModifierStats(target, log, start, end);
+            if (res != null)
+            {
+                return res;
+            }
+            //
+            var damageMods = new List<DamageModifier>();
+            if (log.DamageModifiers.DamageModifiersPerSource.TryGetValue(Source.Item, out IReadOnlyList<DamageModifier> list))
+            {
+                damageMods.AddRange(list);
+            }
+            if (log.DamageModifiers.DamageModifiersPerSource.TryGetValue(Source.Gear, out list))
+            {
+                damageMods.AddRange(list);
+            }
+            if (log.DamageModifiers.DamageModifiersPerSource.TryGetValue(Source.Common, out list))
+            {
+                damageMods.AddRange(list);
+            }
+            if (log.DamageModifiers.DamageModifiersPerSource.TryGetValue(Source.FightSpecific, out list))
+            {
+                damageMods.AddRange(list);
+            }
+            damageMods.AddRange(log.DamageModifiers.GetModifiersPerProf(Prof));
+            //
+            var damageModifierEvents = new List<DamageModifierEvent>();
+            foreach (DamageModifier damageMod in damageMods)
+            {
+                damageModifierEvents.AddRange(damageMod.ComputeDamageModifier(this, log));
+            }
+            damageModifierEvents.Sort((x, y) => x.Time.CompareTo(y.Time));
+            var damageModifiersEvents = damageModifierEvents.GroupBy(y => y.DamageModifier.Name).ToDictionary(y => y.Key, y => y.ToList());
+            _damageModifierEventsPerTargets.Set(log.FightData.FightStart, log.FightData.FightEnd, null, damageModifiersEvents);
+            var damageModifiersEventsByTarget = damageModifierEvents.GroupBy(x => x.Dst).ToDictionary(x => x.Key, x => x.GroupBy(y => y.DamageModifier.Name).ToDictionary(y => y.Key, y => y.ToList()));
+            foreach (AgentItem actor in damageModifiersEventsByTarget.Keys)
+            {
+                _damageModifierEventsPerTargets.Set(log.FightData.FightStart, log.FightData.FightEnd, log.FindActor(actor), damageModifiersEventsByTarget[actor]);
+            }
+            //
+            res = ComputeDamageModifierStats(target, log, start, end);
+            return res;
+        }
+
+        public IReadOnlyCollection<string> GetPresentDamageModifier(ParsedEvtcLog log)
+        {
+            return new HashSet<string>(GetDamageModifierStats(null, log, log.FightData.FightStart, log.FightData.FightEnd).Keys);
+        }
+
         // Buffs
         public BuffDistribution GetBuffDistribution(ParsedEvtcLog log, long start, long end)
         {
@@ -377,48 +490,50 @@ namespace GW2EIEvtcParser.EIData
             }
         }
 
-        public double GetCurrentHealthPercent(ParsedEvtcLog log, long time)
+        public IReadOnlyDictionary<long, FinalActorBuffs> GetBuffs(BuffEnum type, ParsedEvtcLog log, long start, long end)
         {
-            IReadOnlyList<Segment> hps = GetHealthUpdates(log);
-            if (!hps.Any())
+            if (_buffStats == null)
             {
-                return -1.0;
+                _buffStats = new CachingCollectionCustom<BuffEnum, Dictionary<long, FinalActorBuffs>[]>(log, BuffEnum.Self);
             }
-            foreach (Segment seg in hps)
+            if (!_buffStats.TryGetValue(start, end, type, out Dictionary<long, FinalActorBuffs>[] value))
             {
-                if (seg.Intersect(time - ParserHelper.ServerDelayConstant, time + ParserHelper.ServerDelayConstant))
-                {
-                    return seg.Value;
-                }
+                value = SetBuffs(log, start, end, type);
+                _buffStats.Set(start, end, type, value);
             }
-            return -1.0;
+            return value[0];
         }
 
-        public double GetCurrentBarrierPercent(ParsedEvtcLog log, long time)
+        public IReadOnlyDictionary<long, FinalActorBuffs> GetActiveBuffs(BuffEnum type, ParsedEvtcLog log, long start, long end)
         {
-            IReadOnlyList<Segment> hps = GetBarrierUpdates(log);
-            if (!hps.Any())
+            if (_buffStats == null)
             {
-                return -1.0;
+                _buffStats = new CachingCollectionCustom<BuffEnum, Dictionary<long, FinalActorBuffs>[]>(log, BuffEnum.Self);
             }
-            foreach (Segment seg in hps)
+            if (!_buffStats.TryGetValue(start, end, type, out Dictionary<long, FinalActorBuffs>[] value))
             {
-                if (seg.Intersect(time - ParserHelper.ServerDelayConstant, time + ParserHelper.ServerDelayConstant))
-                {
-                    return seg.Value;
-                }
+                value = SetBuffs(log, start, end, type);
+                _buffStats.Set(start, end, type, value);
             }
-            return -1.0;
+            return value[1];
         }
 
-        public Point3D GetCurrentPosition(ParsedEvtcLog log, long time)
+
+        protected virtual Dictionary<long, FinalActorBuffs>[] SetBuffs(ParsedEvtcLog log, long start, long end, BuffEnum type)
         {
-            IReadOnlyList<Point3D> positions = GetCombatReplayPolledPositions(log);
-            if (!positions.Any())
+            switch (type)
             {
-                return null;
+                case BuffEnum.Group:
+                    return null;
+                case BuffEnum.OffGroup:
+                    return null;
+                case BuffEnum.Squad:
+                    var otherPlayers = log.PlayerList.Where(p => p.Agent != Agent).ToList();
+                    return FinalActorBuffs.GetBuffsForPlayers(otherPlayers, log, AgentItem, start, end);
+                case BuffEnum.Self:
+                default:
+                    return FinalActorBuffs.GetBuffsForSelf(log, this, start, end);
             }
-            return positions.FirstOrDefault(x => x.Time >= time);
         }
 
         public IReadOnlyCollection<Buff> GetTrackedBuffs(ParsedEvtcLog log)
@@ -912,5 +1027,50 @@ namespace GW2EIEvtcParser.EIData
             }
             return dls;
         }
+
+        public double GetCurrentHealthPercent(ParsedEvtcLog log, long time)
+        {
+            IReadOnlyList<Segment> hps = GetHealthUpdates(log);
+            if (!hps.Any())
+            {
+                return -1.0;
+            }
+            foreach (Segment seg in hps)
+            {
+                if (seg.Intersect(time - ParserHelper.ServerDelayConstant, time + ParserHelper.ServerDelayConstant))
+                {
+                    return seg.Value;
+                }
+            }
+            return -1.0;
+        }
+
+        public double GetCurrentBarrierPercent(ParsedEvtcLog log, long time)
+        {
+            IReadOnlyList<Segment> hps = GetBarrierUpdates(log);
+            if (!hps.Any())
+            {
+                return -1.0;
+            }
+            foreach (Segment seg in hps)
+            {
+                if (seg.Intersect(time - ParserHelper.ServerDelayConstant, time + ParserHelper.ServerDelayConstant))
+                {
+                    return seg.Value;
+                }
+            }
+            return -1.0;
+        }
+
+        public Point3D GetCurrentPosition(ParsedEvtcLog log, long time)
+        {
+            IReadOnlyList<Point3D> positions = GetCombatReplayPolledPositions(log);
+            if (!positions.Any())
+            {
+                return null;
+            }
+            return positions.FirstOrDefault(x => x.Time >= time);
+        }
+
     }
 }
